move env function to env module
